+++ test/NuGet.Core.Tests/NuGet.Packaging.Test/ManifestTest.cs
@@ -10,6 +10,7 @@ using System.Xml.Linq;
 using NuGet.Frameworks;
 using NuGet.Packaging.Core;
 using NuGet.Versioning;
+using Test.Utility;
 using Xunit;
 
 namespace NuGet.Packaging.Test
