+++ test/NuGet.Core.Tests/NuGet.Packaging.Test/PackageBuilderTest.cs
@@ -24,6 +24,7 @@ using NuGet.Packaging.PackageCreation.Resources;
 using NuGet.ProjectModel;
 using NuGet.Test.Utility;
 using NuGet.Versioning;
+using Test.Utility;
 using Xunit;
 
 namespace NuGet.Packaging.Test
